Feature/reorganize code (#39)
* restart exited container on device connection

* fix not working wda stream settings on new container

* remove non-working caps from ios sync script

* rework container creation function a bit

* small changes and cleanup

* change response struct naming a bit

* more changes

* fix project config json field ref

* fix listener endpoints in web code

* move code around and cleanup a bit

* more code restructure

* add comments and small reworks

* more comments and updates

* more updates

* update setup docs

* update docs

* change viewport size to screen size
