+++ devices.go
@@ -2,32 +2,29 @@ package main
 
 import (
 	"bytes"
+	"context"
 	"errors"
 	"fmt"
+	"html/template"
 	"io/ioutil"
 	"net/http"
 	"os/exec"
+	"path/filepath"
+	"regexp"
+	"strconv"
+	"strings"
 
 	"github.com/danielpaulus/go-ios/ios"
 	"github.com/danielpaulus/go-ios/ios/installationproxy"
 	"github.com/danielpaulus/go-ios/ios/zipconduit"
+	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/client"
 	"github.com/gorilla/mux"
 	log "github.com/sirupsen/logrus"
 )
 
 //=================//
 //=====STRUCTS=====//
-
-type detailsList struct {
-	DetailsList []detailsEntry `json:"deviceList"`
-}
-
-type detailsEntry struct {
-	Udid           string
-	ProductName    string
-	ProductType    string
-	ProductVersion string
-}
 
 type IOSDeviceInfo struct {
 	BundleIDs    []string   `json:"installedAppsBundleIDs"`
@@ -39,19 +36,11 @@ type AndroidDeviceInfo struct {
 	DeviceConfig *AndroidDevice `json:"deviceConfig"`
 }
 
-type iOSDevice struct {
-	AppiumPort      int    `json:"appium_port"`
-	DeviceName      string `json:"device_name"`
-	DeviceOSVersion string `json:"device_os_version"`
-	DeviceUDID      string `json:"device_udid"`
-	WdaMjpegPort    int    `json:"wda_mjpeg_port"`
-	WdaPort         int    `json:"wda_port"`
-}
-
-type registerIOSDevice struct {
-	DeviceUDID      string `json:"device_udid"`
-	DeviceName      string `json:"device_name"`
-	DeviceOSVersion string `json:"device_os_version"`
+type DeviceControlInfo struct {
+	RunningContainers []string            `json:"running-containers"`
+	IOSInfo           []IOSDeviceInfo     `json:"ios-devices-info"`
+	AndroidInfo       []AndroidDeviceInfo `json:"android-devices-info"`
+	InstallableApps   []string            `json:"installable-apps"`
 }
 
 type installIOSAppRequest struct {
@@ -69,14 +58,43 @@ type goIOSAppList []struct {
 //=======================//
 //=====API FUNCTIONS=====//
 
+// Load the device control page
+func GetDeviceControlPage(w http.ResponseWriter, r *http.Request) {
+	var device_control_page = template.Must(template.ParseFiles("static/device_control.html"))
+	if err := device_control_page.Execute(w, nil); err != nil {
+		log.WithFields(log.Fields{
+			"event": "device_control_page",
+		}).Error("Couldn't load device_control.html")
+		return
+	}
+}
+
+// @Summary      Get device control info
+// @Description  Provides the running containers, IOS devices info and apps available for installing
+// @Tags         devices
+// @Produce      json
+// @Success      200 {object} DeviceControlInfo
+// @Failure      500 {object} JsonErrorResponse
+// @Router       /devices/device-control [post]
+func GetDeviceControlInfo(w http.ResponseWriter, r *http.Request) {
+	var runningContainerNames = getRunningDeviceContainerNames()
+	var info = DeviceControlInfo{
+		RunningContainers: runningContainerNames,
+		IOSInfo:           getIOSDevicesInfo(runningContainerNames),
+		InstallableApps:   getInstallableApps(),
+		AndroidInfo:       getAndroidDevicesInfo(runningContainerNames),
+	}
+	fmt.Fprintf(w, PrettifyJSON(ConvertToJSONString(info)))
+}
+
 // @Summary      Get logs for iOS device container
 // @Description  Get logs by type
 // @Tags         device-logs
 // @Produce      json
 // @Param        log_type path string true "Log Type"
 // @Param        device_udid path string true "Device UDID"
-// @Success      200 {object} SimpleResponseJSON
-// @Failure      200 {object} SimpleResponseJSON
+// @Success      200 {object} JsonResponse
+// @Failure      200 {object} JsonResponse
 // @Router       /device-logs/{log_type}/{device_udid} [get]
 func GetDeviceLogs(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
@@ -101,45 +119,14 @@ func GetDeviceLogs(w http.ResponseWriter, r *http.Request) {
 	SimpleJSONResponse(w, out.String(), 200)
 }
 
-// @Summary      Get connected iOS devices
-// @Description  Returns the connected iOS devices with go-ios
-// @Tags         ios-devices
-// @Produce      json
-// @Success      200 {object} detailsList
-// @Failure      500 {object} ErrorJSON
-// @Router       /ios-devices [get]
-func GetConnectedIOSDevices(w http.ResponseWriter, r *http.Request) {
-	deviceList, err := ios.ListDevices()
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "get_connected_ios_devices",
-		}).Error("Could not get connected devices. Error: " + err.Error())
-		JSONError(w, "get_connected_ios_devices", "Could not get connected devices. Error: "+err.Error(), 500)
-		return
-	}
-	deviceValues, err := outputDetailedList(deviceList)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "get_connected_ios_devices",
-		}).Error("Could not get connected devices detailed list. Error: " + err.Error())
-		JSONError(w, "get_connected_ios_devices", "Could not get connected devices detailed list. Error: "+err.Error(), 500)
-		return
-	}
-	log.WithFields(log.Fields{
-		"event": "get_connected_ios_devices",
-	}).Info("Successfully got connected iOS devices detailed list.")
-
-	fmt.Fprintf(w, PrettifyJSON(deviceValues))
-}
-
 // @Summary      Install app on iOS device
 // @Description  Installs *.ipa or *.app from the './apps' folder with go-ios
 // @Tags         ios-devices
 // @Produce      json
 // @Param        device_udid path string true "Device UDID"
 // @Param        config body installIOSAppRequest true "Install iOS app"
-// @Success      200 {object} SimpleResponseJSON
-// @Failure      500 {object} ErrorJSON
+// @Success      200 {object} JsonResponse
+// @Failure      500 {object} JsonErrorResponse
 // @Router       /ios-devices/{device_udid}/install-app [post]
 func InstallIOSApp(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
@@ -167,8 +154,8 @@ func InstallIOSApp(w http.ResponseWriter, r *http.Request) {
 // @Produce      json
 // @Param        device_udid path string true "Device UDID"
 // @Param        config body uninstallIOSAppRequest true "Uninstall iOS app"
-// @Success      200 {object} SimpleResponseJSON
-// @Failure      500 {object} ErrorJSON
+// @Success      200 {object} JsonResponse
+// @Failure      500 {object} JsonErrorResponse
 // @Router       /ios-devices/{device_udid}/uninstall-app [post]
 func UninstallIOSApp(w http.ResponseWriter, r *http.Request) {
 	var data uninstallIOSAppRequest
@@ -196,21 +183,6 @@ func UninstallIOSApp(w http.ResponseWriter, r *http.Request) {
 
 //===================//
 //=====FUNCTIONS=====//
-
-func outputDetailedList(deviceList ios.DeviceList) (string, error) {
-	result := make([]detailsEntry, len(deviceList.DeviceList))
-	for i, device := range deviceList.DeviceList {
-		udid := device.Properties.SerialNumber
-		allValues, err := ios.GetValues(device)
-		if err != nil {
-			return "", errors.New("Failed getting device values")
-		}
-		result[i] = detailsEntry{udid, allValues.Value.ProductName, allValues.Value.ProductType, allValues.Value.ProductVersion}
-	}
-	return ConvertToJSONString(map[string][]detailsEntry{
-		"deviceList": result,
-	}), nil
-}
 
 func IOSDeviceApps(device_udid string) ([]string, error) {
 	device, err := ios.GetDevice(device_udid)
@@ -315,4 +287,231 @@ func uninstallIOSApp(device_udid string, bundle_id string) error {
 		return errors.New("Error")
 	}
 	return nil
+}
+
+// For each running container extract the info for each respective device from ./configs/config.json to provide to the device-control info endpoint.
+// Provides installed apps, configuration info, wda urls
+func getIOSDevicesInfo(runningContainers []string) []IOSDeviceInfo {
+	var combinedInfo []IOSDeviceInfo
+	for _, containerName := range runningContainers {
+		if strings.Contains(containerName, "iosDevice") {
+			// Extract the device UDID from the container name
+			re := regexp.MustCompile("[^_]*$")
+			device_udid := re.FindStringSubmatch(containerName)
+
+			var installed_apps []string
+			installed_apps, err := IOSDeviceApps(device_udid[0])
+			if err != nil {
+				installed_apps = append(installed_apps, "")
+			}
+
+			var device_config *IOSDevice
+			device_config, err = iOSDeviceConfig(device_udid[0])
+
+			var deviceInfo = IOSDeviceInfo{BundleIDs: installed_apps, DeviceConfig: device_config}
+			combinedInfo = append(combinedInfo, deviceInfo)
+		}
+	}
+	return combinedInfo
+}
+
+// For each running container extract the info for each respective device from ./configs/config.json to provide to the device-control info endpoint.
+// Provides installed apps, configuration info, wda urls
+func getAndroidDevicesInfo(runningContainers []string) []AndroidDeviceInfo {
+	var combinedInfo []AndroidDeviceInfo
+	for _, containerName := range runningContainers {
+		if strings.Contains(containerName, "androidDevice") {
+			// Extract the device UDID from the container name
+			re := regexp.MustCompile("[^_]*$")
+			device_udid := re.FindStringSubmatch(containerName)
+
+			var device_config *AndroidDevice
+			device_config, _ = androidDeviceConfig(device_udid[0])
+
+			var deviceInfo = AndroidDeviceInfo{DeviceConfig: device_config}
+			combinedInfo = append(combinedInfo, deviceInfo)
+		}
+	}
+	return combinedInfo
+}
+
+// Get the configuration info for iOS device from ./configs/config.json
+func iOSDeviceConfig(device_udid string) (*IOSDevice, error) {
+	// Get the config data
+	configData, err := GetConfigJsonData()
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "ios_container_create",
+		}).Error("Could not unmarshal config.json file when trying to create a container for device with udid: " + device_udid)
+		return nil, err
+	}
+
+	// Check if device is registered in config data
+	var device_in_config bool
+	var deviceConfig DeviceConfig
+	for _, v := range configData.DeviceConfig {
+		if v.DeviceUDID == device_udid {
+			device_in_config = true
+			deviceConfig = v
+		}
+	}
+
+	// Stop execution if device not in config data
+	if !device_in_config {
+		log.WithFields(log.Fields{
+			"event": "android_container_create",
+		}).Error("Device with UDID:" + device_udid + " is not registered in the './configs/config.json' file. No container will be created.")
+		return nil, err
+	}
+
+	wda_url, wda_stream_url, err := getIOSDeviceWdaURLs(device_udid)
+	if err != nil {
+		return nil, err
+	}
+
+	return &IOSDevice{
+			AppiumPort:       deviceConfig.AppiumPort,
+			DeviceName:       deviceConfig.DeviceName,
+			DeviceOSVersion:  deviceConfig.DeviceOSVersion,
+			WdaMjpegPort:     deviceConfig.WDAMjpegPort,
+			WdaPort:          deviceConfig.WDAPort,
+			DeviceUDID:       device_udid,
+			WdaMjpegURL:      wda_stream_url,
+			WdaURL:           wda_url,
+			DeviceModel:      "Remove please",
+			DeviceScreenSize: deviceConfig.ScreenSize},
+		nil
+}
+
+// Get the configuration info for iOS device from ./configs/config.json
+func androidDeviceConfig(device_udid string) (*AndroidDevice, error) {
+	// Get the config data
+	configData, err := GetConfigJsonData()
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "android_container_create",
+		}).Error("Could not unmarshal config.json file when trying to create a container for device with udid: " + device_udid)
+		return nil, err
+	}
+
+	// Check if device is registered in config data
+	var device_in_config bool
+	var deviceConfig DeviceConfig
+	for _, v := range configData.DeviceConfig {
+		if v.DeviceUDID == device_udid {
+			device_in_config = true
+			deviceConfig = v
+		}
+	}
+
+	// Stop execution if device not in config data
+	if !device_in_config {
+		log.WithFields(log.Fields{
+			"event": "android_container_create",
+		}).Error("Device with UDID:" + device_udid + " is not registered in the './configs/config.json' file. No container will be created.")
+		return nil, err
+	}
+
+	stream_size, err := getAndroidDeviceMinicapStreamSize(device_udid)
+	return &AndroidDevice{
+			AppiumPort:      deviceConfig.AppiumPort,
+			DeviceName:      deviceConfig.DeviceName,
+			DeviceOSVersion: deviceConfig.DeviceOSVersion,
+			DeviceUDID:      device_udid,
+			StreamSize:      stream_size,
+			StreamPort:      strconv.Itoa(deviceConfig.StreamPort)},
+		nil
+}
+
+// Get the installable apps list from the ./apps folder
+func getInstallableApps() []string {
+	var appNames []string
+
+	files, err := ioutil.ReadDir("./apps/")
+	if err != nil {
+		return appNames
+	}
+
+	for _, file := range files {
+		if strings.Contains(file.Name(), ".ipa") || strings.Contains(file.Name(), ".app") || strings.Contains(file.Name(), ".apk") {
+			appNames = append(appNames, file.Name())
+		}
+	}
+	return appNames
+}
+
+// Get the WDA and WDA stream urls from the container logs folder for a specific device
+func getIOSDeviceWdaURLs(device_udid string) (string, string, error) {
+	// Get the path of the WDA url file using regex
+	pattern := "./logs/*" + device_udid + "/ios-wda-url.json"
+	matches, err := filepath.Glob(pattern)
+
+	if err != nil {
+		return "", "", err
+	}
+
+	var wdaConfig WdaConfig
+	err = UnmarshalJSONFile(matches[0], &wdaConfig)
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "get_wda_config",
+		}).Error("Could not unmarshal json at path: " + matches[0])
+		return "", "", err
+	}
+
+	return wdaConfig.WdaURL, wdaConfig.WdaStreamURL, nil
+}
+
+// Get the WDA and WDA stream urls from the container logs folder for a specific device
+func getAndroidDeviceMinicapStreamSize(device_udid string) (string, error) {
+	// Get the path of the WDA url file using regex
+	pattern := "./logs/*" + device_udid + "/minicap.log"
+	matches, err := filepath.Glob(pattern)
+
+	command_string := "cat " + matches[0] + " | grep \"+ args='-P\""
+	cmd := exec.Command("bash", "-c", command_string)
+	var out bytes.Buffer
+	cmd.Stdout = &out
+
+	err = cmd.Run()
+	if err != nil {
+		return "", err
+	}
+
+	if out.String() == "" {
+		log.WithFields(log.Fields{
+			"event": "get_minicap_stream_size",
+		}).Error("Error")
+		return "", nil
+	}
+
+	// Get the stream size that is between the -P and @ in the cmd out string
+	stream_size := GetStringInBetween(out.String(), "-P ", "@")
+	return stream_size, nil
+}
+
+// Get the names of the currently running containers(that are for devices)
+func getRunningDeviceContainerNames() []string {
+	var containerNames []string
+
+	cli, err := client.NewClientWithOpts(client.FromEnv)
+	if err != nil {
+		return containerNames
+	}
+
+	// Get the current containers list
+	containers, err := cli.ContainerList(context.Background(), types.ContainerListOptions{All: true})
+	if err != nil {
+		return containerNames
+	}
+
+	// Loop through the containers list
+	for _, container := range containers {
+		// Parse plain container name
+		containerName := strings.Replace(container.Names[0], "/", "", -1)
+		if (strings.Contains(containerName, "iosDevice") || strings.Contains(containerName, "androidDevice")) && strings.Contains(container.Status, "Up") {
+			containerNames = append(containerNames, containerName)
+		}
+	}
+	return containerNames
 }
