+++ main.go
@@ -1,32 +1,18 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
-	"fmt"
 	"html/template"
-	"io/ioutil"
 	"net/http"
 	"os"
-	"os/exec"
 
 	_ "GADS/docs"
 
 	"github.com/gorilla/mux"
-	"github.com/gorilla/websocket"
 	log "github.com/sirupsen/logrus"
 	httpSwagger "github.com/swaggo/http-swagger"
 )
 
 var project_log_file *os.File
-
-type ProjectConfigPageData struct {
-	WebDriverAgentProvided bool
-	SudoPasswordSet        bool
-	UdevIOSListenerStatus  string
-	ImageStatus            string
-	ProjectConfigValues    AppiumConfig
-}
 
 type AppiumConfig struct {
 	DevicesHost             string `json:"devices_host"`
@@ -44,7 +30,7 @@ type DeviceConfig struct {
 	DeviceUDID      string `json:"device_udid"`
 	WDAMjpegPort    int    `json:"wda_mjpeg_port,omitempty"`
 	WDAPort         int    `json:"wda_port,omitempty"`
-	ViewportSize    string `json:"viewport_size,omitempty"`
+	ScreenSize      string `json:"screen_size,omitempty"`
 	StreamPort      int    `json:"stream_port,omitempty"`
 }
 
@@ -71,80 +57,6 @@ func GetInitialPage(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// Load the initial page with the project configuration info
-func GetProjectConfigurationPage(w http.ResponseWriter, r *http.Request) {
-	projectConfig, err := GetConfigJsonData()
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	var configRow = AppiumConfig{
-		DevicesHost:             projectConfig.AppiumConfig.DevicesHost,
-		SeleniumHubHost:         projectConfig.AppiumConfig.SeleniumHubHost,
-		SeleniumHubPort:         projectConfig.AppiumConfig.SeleniumHubPort,
-		SeleniumHubProtocolType: projectConfig.AppiumConfig.SeleniumHubProtocolType,
-		WDABundleID:             projectConfig.AppiumConfig.WDABundleID}
-
-	var index = template.Must(template.ParseFiles("static/project_config.html"))
-	pageData := ProjectConfigPageData{WebDriverAgentProvided: CheckWDAProvided(), SudoPasswordSet: CheckSudoPasswordSet(), UdevIOSListenerStatus: UdevIOSListenerState(), ImageStatus: ImageExists(), ProjectConfigValues: configRow}
-	if err := index.Execute(w, pageData); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
-}
-
-// @Summary      Update project configuration
-// @Description  Updates one  or multiple configuration values
-// @Tags         configuration
-// @Param        config body ProjectConfig true "Update config"
-// @Accept		 json
-// @Produce      json
-// @Success      200 {object} SimpleResponseJSON
-// @Failure      500 {object} ErrorJSON
-// @Router       /configuration/update-config [put]
-func UpdateProjectConfigHandler(w http.ResponseWriter, r *http.Request) {
-	var requestData AppiumConfig
-	err := UnmarshalRequestBody(r.Body, &requestData)
-
-	// Get the config data
-	configData, err := GetConfigJsonData()
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "ios_container_create",
-		}).Error("Could not unmarshal config.json file when trying to change configuration values.")
-		return
-	}
-
-	if requestData.DevicesHost != "" {
-		configData.AppiumConfig.DevicesHost = requestData.DevicesHost
-	}
-	if requestData.SeleniumHubHost != "" {
-		configData.AppiumConfig.SeleniumHubHost = requestData.SeleniumHubHost
-	}
-	if requestData.SeleniumHubPort != "" {
-		configData.AppiumConfig.SeleniumHubPort = requestData.SeleniumHubPort
-	}
-	if requestData.SeleniumHubProtocolType != "" {
-		configData.AppiumConfig.SeleniumHubProtocolType = requestData.SeleniumHubProtocolType
-	}
-	if requestData.WDABundleID != "" {
-		configData.AppiumConfig.WDABundleID = requestData.WDABundleID
-	}
-
-	bs, err := json.MarshalIndent(configData, "", "  ")
-
-	err = ioutil.WriteFile("./configs/config.json", bs, 0644)
-	if err != nil {
-		JSONError(w, "config_file_interaction", "Could not write to the config.json file.", 500)
-		return
-	}
-	SimpleJSONResponse(w, "Successfully updated project config in ./configs/config.json", 200)
-}
-
-var upgrader = websocket.Upgrader{
-	ReadBufferSize:  1024,
-	WriteBufferSize: 1024,
-}
-
 func setLogging() {
 	log.SetFormatter(&log.JSONFormatter{})
 	project_log_file, err := os.OpenFile("./logs/project.log", os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0755)
@@ -152,55 +64,6 @@ func setLogging() {
 		panic(err)
 	}
 	log.SetOutput(project_log_file)
-}
-
-func GetLogsPage(w http.ResponseWriter, r *http.Request) {
-	var logs_page = template.Must(template.ParseFiles("static/project_logs.html"))
-	if err := logs_page.Execute(w, nil); err != nil {
-		log.WithFields(log.Fields{
-			"event": "project_logs_page",
-		}).Error("Couldn't load project_logs.html")
-		return
-	}
-}
-
-func GetDeviceControlPage(w http.ResponseWriter, r *http.Request) {
-	var device_control_page = template.Must(template.ParseFiles("static/device_control.html"))
-	if err := device_control_page.Execute(w, nil); err != nil {
-		log.WithFields(log.Fields{
-			"event": "device_control_page",
-		}).Error("Couldn't load device_control.html")
-		return
-	}
-}
-
-// @Summary      Get project logs
-// @Description  Provides project logs as plain text response
-// @Tags         project-logs
-// @Success      200
-// @Failure      200
-// @Router       /project-logs [get]
-func GetLogs(w http.ResponseWriter, r *http.Request) {
-	// Execute the command to restart the container by container ID
-	commandString := "tail -n 1000 ./logs/project.log"
-	cmd := exec.Command("bash", "-c", commandString)
-	var out bytes.Buffer
-	cmd.Stdout = &out
-	err := cmd.Run()
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "get_project_logs",
-		}).Error("Attempted to get project logs but no logs available.")
-		fmt.Fprintf(w, "No logs available")
-		return
-	}
-	//SimpleJSONResponse(w, "get_project_logs", out.String(), 200)
-	fmt.Fprintf(w, out.String())
-}
-
-type Order struct {
-	Username string `json:"username"`
-	Fullname string `json:"fullname"`
 }
 
 func handleRequests() {
@@ -216,10 +79,6 @@ func handleRequests() {
 		httpSwagger.DomID("#swagger-ui"),
 	))
 
-	// iOS containers endpoints
-
-	// Android containers endpoints
-
 	// General containers endpoints
 	myRouter.HandleFunc("/containers/{container_id}/restart", RestartContainer).Methods("POST")
 	myRouter.HandleFunc("/containers/{container_id}/remove", RemoveContainer).Methods("POST")
@@ -229,9 +88,9 @@ func handleRequests() {
 
 	// Configuration endpoints
 	myRouter.HandleFunc("/configuration/build-image/{image_type}", BuildDockerImage).Methods("POST")
-	myRouter.HandleFunc("/configuration/remove-image", RemoveDockerImage).Methods("POST")
-	myRouter.HandleFunc("/configuration/setup-ios-listener", SetupUdevListener).Methods("POST")
-	myRouter.HandleFunc("/configuration/remove-ios-listener", RemoveUdevListener).Methods("POST")
+	myRouter.HandleFunc("/configuration/remove-image/{image_type}", RemoveDockerImage).Methods("POST")
+	myRouter.HandleFunc("/configuration/setup-udev-listener", SetupUdevListener).Methods("POST")
+	myRouter.HandleFunc("/configuration/remove-udev-listener", RemoveUdevListener).Methods("POST")
 	myRouter.HandleFunc("/configuration/update-config", UpdateProjectConfigHandler).Methods("PUT")
 	myRouter.HandleFunc("/configuration/set-sudo-password", SetSudoPassword).Methods("PUT")
 	myRouter.HandleFunc("/configuration/upload-wda", UploadWDA).Methods("POST")
@@ -239,7 +98,6 @@ func handleRequests() {
 
 	// Devices endpoints
 	myRouter.HandleFunc("/device-logs/{log_type}/{device_udid}", GetDeviceLogs).Methods("GET")
-	myRouter.HandleFunc("/ios-devices", GetConnectedIOSDevices).Methods("GET")
 	myRouter.HandleFunc("/ios-devices/{device_udid}/install-app", InstallIOSApp).Methods("POST")
 	myRouter.HandleFunc("/ios-devices/{device_udid}/uninstall-app", UninstallIOSApp).Methods("POST")
 	myRouter.HandleFunc("/devices/device-control", GetDeviceControlInfo).Methods("GET")
