+++ project_control.go
@@ -2,21 +2,16 @@ package main
 
 import (
 	"bytes"
-	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
+	"html/template"
 	"io/ioutil"
 	"net/http"
 	"os"
 	"os/exec"
-	"path/filepath"
-	"regexp"
 	"strconv"
-	"strings"
 
-	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/client"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -25,28 +20,17 @@ var sudo_password = GetEnvValue("sudo_password")
 //=================//
 //=====STRUCTS=====//
 
-type SudoPassword struct {
-	SudoPassword string `json:"sudo_password"`
-}
-
-type DeviceControlInfo struct {
-	RunningContainers []string            `json:"running-containers"`
-	IOSInfo           []IOSDeviceInfo     `json:"ios-devices-info"`
-	AndroidInfo       []AndroidDeviceInfo `json:"android-devices-info"`
-	InstallableApps   []string            `json:"installable-apps"`
-}
-
 type IOSDevice struct {
-	AppiumPort         int    `json:"appium_port"`
-	DeviceName         string `json:"device_name"`
-	DeviceOSVersion    string `json:"device_os_version"`
-	DeviceUDID         string `json:"device_udid"`
-	WdaMjpegPort       int    `json:"wda_mjpeg_port"`
-	WdaPort            int    `json:"wda_port"`
-	WdaURL             string `json:"wda_url"`
-	WdaMjpegURL        string `json:"wda_stream_url"`
-	DeviceModel        string `json:"device_model"`
-	DeviceViewportSize string `json:"viewport_size"`
+	AppiumPort       int    `json:"appium_port"`
+	DeviceName       string `json:"device_name"`
+	DeviceOSVersion  string `json:"device_os_version"`
+	DeviceUDID       string `json:"device_udid"`
+	WdaMjpegPort     int    `json:"wda_mjpeg_port"`
+	WdaPort          int    `json:"wda_port"`
+	WdaURL           string `json:"wda_url"`
+	WdaMjpegURL      string `json:"wda_stream_url"`
+	DeviceModel      string `json:"device_model"`
+	DeviceScreenSize string `json:"screen_size"`
 }
 
 type AndroidDevice struct {
@@ -58,53 +42,171 @@ type AndroidDevice struct {
 	StreamPort      string `json:"stream_port"`
 }
 
-type IOSDeviceConfig struct {
-	AppiumPort      int    `json:"appium_port"`
-	DeviceName      string `json:"device_name"`
-	DeviceOsVersion string `json:"device_os_version"`
-	DeviceUdid      string `json:"device_udid"`
-	WdaMjpegPort    int    `json:"wda_mjpeg_port"`
-	WdaPort         int    `json:"wda_port"`
-}
-
-type AndroidDeviceConfig struct {
-	AppiumPort      int    `json:"appium_port"`
-	DeviceUdid      string `json:"device_udid"`
-	DeviceName      string `json:"device_name"`
-	StreamPort      int    `json:"stream_port"`
-	DeviceOsVersion string `json:"device_os_version"`
-}
-
 type WdaConfig struct {
 	WdaURL       string `json:"wda_url"`
 	WdaStreamURL string `json:"wda_stream_url"`
-}
-
-type GoIOSDevicesList struct {
-	DeviceList []GoIOSDevice `json:"deviceList"`
-}
-
-type GoIOSDevice struct {
-	Udid           string `json:"Udid"`
-	ProductName    string `json:"ProductName"`
-	ProductType    string `json:"ProductType"`
-	ProductVersion string `json:"ProductVersion"`
 }
 
 type SudoPasswordRequest struct {
 	SudoPassword string `json:"sudo_password"`
 }
 
+type ProjectConfigPageData struct {
+	WebDriverAgentProvided bool
+	SudoPasswordSet        bool
+	UdevIOSListenerStatus  string
+	ImageStatus            string
+	AppiumConfigValues     AppiumConfig
+}
+
 //=======================//
 //=====API FUNCTIONS=====//
 
-// @Summary      Sets up iOS device listener
+// Load the general logs page
+func GetLogsPage(w http.ResponseWriter, r *http.Request) {
+	var logs_page = template.Must(template.ParseFiles("static/project_logs.html"))
+	if err := logs_page.Execute(w, nil); err != nil {
+		log.WithFields(log.Fields{
+			"event": "project_logs_page",
+		}).Error("Couldn't load project_logs.html")
+		return
+	}
+}
+
+// Load the page with the project configuration info
+func GetProjectConfigurationPage(w http.ResponseWriter, r *http.Request) {
+	// Get the config data
+	projectConfig, err := GetConfigJsonData()
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	// Create the AppiumConfig
+	var configRow = AppiumConfig{
+		DevicesHost:             projectConfig.AppiumConfig.DevicesHost,
+		SeleniumHubHost:         projectConfig.AppiumConfig.SeleniumHubHost,
+		SeleniumHubPort:         projectConfig.AppiumConfig.SeleniumHubPort,
+		SeleniumHubProtocolType: projectConfig.AppiumConfig.SeleniumHubProtocolType,
+		WDABundleID:             projectConfig.AppiumConfig.WDABundleID}
+
+	var index = template.Must(template.ParseFiles("static/project_config.html"))
+
+	// Create the final data for the config page
+	pageData := ProjectConfigPageData{WebDriverAgentProvided: CheckWDAProvided(), SudoPasswordSet: CheckSudoPasswordSet(), UdevIOSListenerStatus: UdevIOSListenerState(), ImageStatus: ImageExists(), AppiumConfigValues: configRow}
+	if err := index.Execute(w, pageData); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+}
+
+// @Summary      Get project logs
+// @Description  Provides project logs as plain text response
+// @Tags         project-logs
+// @Produces	 text
+// @Success      200
+// @Failure      200
+// @Router       /project-logs [get]
+func GetLogs(w http.ResponseWriter, r *http.Request) {
+	// Create the command string to read the last 1000 lines of project.log
+	commandString := "tail -n 1000 ./logs/project.log"
+
+	// Create the command
+	cmd := exec.Command("bash", "-c", commandString)
+
+	// Create a buffer for the output
+	var out bytes.Buffer
+
+	// Pipe the Stdout of the command to the buffer pointer
+	cmd.Stdout = &out
+
+	// Execute the command
+	err := cmd.Run()
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "get_project_logs",
+		}).Error("Attempted to get project logs but no logs available.")
+
+		// Reply with generic message on error
+		fmt.Fprintf(w, "No logs available")
+		return
+	}
+
+	// Reply with the read logs lines
+	fmt.Fprintf(w, out.String())
+}
+
+// @Summary      Update project Appium configuration
+// @Description  Updates one or multiple configuration values
+// @Tags         configuration
+// @Param        config body AppiumConfig true "Update config"
+// @Accept		 json
+// @Produce      json
+// @Success      200 {object} JsonResponse
+// @Failure      500 {object} JsonErrorResponse
+// @Router       /configuration/update-config [put]
+func UpdateProjectConfigHandler(w http.ResponseWriter, r *http.Request) {
+	var requestData AppiumConfig
+	// Get the request data
+	err := UnmarshalRequestBody(r.Body, &requestData)
+
+	// Get the config data from configs/config.json
+	configData, err := GetConfigJsonData()
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "update_appium_config",
+		}).Error("Could not get config data: " + err.Error())
+		JSONError(w, "update_appium_config", "An error occurred", 500)
+		return
+	}
+
+	// Check each field from the request
+	// If non-empty value is received update the respective value in the configData
+	if requestData.DevicesHost != "" {
+		configData.AppiumConfig.DevicesHost = requestData.DevicesHost
+	}
+	if requestData.SeleniumHubHost != "" {
+		configData.AppiumConfig.SeleniumHubHost = requestData.SeleniumHubHost
+	}
+	if requestData.SeleniumHubPort != "" {
+		configData.AppiumConfig.SeleniumHubPort = requestData.SeleniumHubPort
+	}
+	if requestData.SeleniumHubProtocolType != "" {
+		configData.AppiumConfig.SeleniumHubProtocolType = requestData.SeleniumHubProtocolType
+	}
+	if requestData.WDABundleID != "" {
+		configData.AppiumConfig.WDABundleID = requestData.WDABundleID
+	}
+
+	// Marshal back the configData into a byte slice
+	bs, err := json.MarshalIndent(configData, "", "  ")
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "update_appium_config",
+		}).Error("Could not marshal back config data: " + err.Error())
+		JSONError(w, "update_appium_config", "An error occurred", 500)
+		return
+	}
+
+	// Write the updated configData json into the file
+	err = ioutil.WriteFile("./configs/config.json", bs, 0644)
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "update_appium_config",
+		}).Error("Could not write back config data: " + err.Error())
+		JSONError(w, "update_appium_config", "An error occurred", 500)
+		return
+	}
+
+	SimpleJSONResponse(w, "Successfully updated project config in ./configs/config.json", 200)
+}
+
+// @Summary      Sets up udev devices listener
 // @Description  Creates udev rules, moves them to /etc/udev/rules.d and reloads udev. Copies usbmuxd.service to /lib/systemd/system and enables it
 // @Tags         configuration
 // @Produce      json
-// @Success      200 {object} SimpleResponseJSON
-// @Failure      500 {object} ErrorJSON
-// @Router       /configuration/setup-ios-listener [post]
+// @Success      200 {object} JsonResponse
+// @Failure      500 {object} JsonErrorResponse
+// @Router       /configuration/setup-udev-listener [post]
 func SetupUdevListener(w http.ResponseWriter, r *http.Request) {
 	// Open /lib/systemd/system/systemd-udevd.service
 	// Add IPAddressAllow=127.0.0.1 at the bottom
@@ -116,67 +218,25 @@ func SetupUdevListener(w http.ResponseWriter, r *http.Request) {
 		JSONError(w, "setup_udev_listener", "Elevated permissions are required to perform this action.", 500)
 		return
 	}
-	DeleteTempUdevFiles()
-	err := CreateUdevRules()
+	err := SetupUdevListenerInternal()
 	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "setup_udev_listener",
-		}).Error("Could not create udev rules file. Error:" + err.Error())
-		JSONError(w, "setup_udev_listener", "Could not create udev rules", 500)
-		DeleteTempUdevFiles()
-		return
+		JSONError(w, "setup_udev_listener", "Could not setup udev rules", 500)
 	}
 
-	err = SetUdevRules()
-	if err != nil {
-		JSONError(w, "setup_udev_listener", err.Error(), 500)
-		DeleteTempUdevFiles()
-		return
-	}
-
-	// err = CopyFileShell("./configs/usbmuxd.service", "/lib/systemd/system/", sudo_password)
-	// if err != nil {
-	// 	DeleteTempUdevFiles()
-	// 	JSONError(w, "setup_udev_listener", "Could not copy usbmuxd.service. Error: "+err.Error(), 500)
-	// 	return
-	// }
-
-	// err = EnableUsbmuxdService()
-	// if err != nil {
-	// 	DeleteTempUdevFiles()
-	// 	JSONError(w, "setup_udev_listener", "Could not enable usbmuxd.service. Error: "+err.Error(), 500)
-	// 	return
-	// }
-
-	DeleteTempUdevFiles()
 	SimpleJSONResponse(w, "Successfully set udev rules.", 200)
 }
 
-// @Summary      Removes iOS device listener
+// @Summary      Removes udev device listener
 // @Description  Deletes udev rules from /etc/udev/rules.d and reloads udev
 // @Tags         configuration
 // @Produce      json
-// @Success      200 {object} SimpleResponseJSON
-// @Failure      500 {object} ErrorJSON
-// @Router       /configuration/remove-ios-listener [post]
+// @Success      200 {object} JsonResponse
+// @Failure      500 {object} JsonErrorResponse
+// @Router       /configuration/remove-device-listener [post]
 func RemoveUdevListener(w http.ResponseWriter, r *http.Request) {
-	err := DeleteFileShell("/etc/udev/rules.d/90-device.rules", sudo_password)
+	err := RemoveUdevListenerInternal()
 	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "remove_udev_listener",
-		}).Error("Could not delete udev rules file. Error: " + err.Error())
-		JSONError(w, "remove_udev_listener", "Could not delete usbmuxd rules file.", 500)
-		return
-	}
-	commandString := "echo '" + sudo_password + "' | sudo -S udevadm control --reload-rules"
-	cmd := exec.Command("bash", "-c", commandString)
-	err = cmd.Run()
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "remove_udev_listener",
-		}).Error("Could not reload udev rules file. Error: " + err.Error())
-		JSONError(w, "remove_udev_listener", "Could not reload udev rules", 500)
-		return
+		JSONError(w, "remove_udev_listener", err.Error(), 500)
 	}
 }
 
@@ -185,13 +245,14 @@ func RemoveUdevListener(w http.ResponseWriter, r *http.Request) {
 // @Tags         configuration
 // @Accept		 json
 // @Produce      json
-// @Param        config body SudoPassword true "Sudo password value"
-// @Success      200 {object} SimpleResponseJSON
-// @Failure      500 {object} ErrorJSON
+// @Param        config body SudoPasswordRequest true "Sudo password value"
+// @Success      200 {object} JsonResponse
+// @Failure      500 {object} JsonErrorResponse
 // @Router       /configuration/set-sudo-password [put]
 func SetSudoPassword(w http.ResponseWriter, r *http.Request) {
 	var requestData SudoPasswordRequest
 
+	// Get the request data
 	err := UnmarshalRequestBody(r.Body, &requestData)
 	if err != nil {
 		log.WithFields(log.Fields{
@@ -211,8 +272,10 @@ func SetSudoPassword(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	// Update the password in configData
 	configData.EnvConfig.SudoPassword = sudo_password
 
+	// Create a byte slice with the updated data
 	bs, err := json.MarshalIndent(configData, "", "  ")
 
 	// Write the new json to the config.json file
@@ -224,32 +287,73 @@ func SetSudoPassword(w http.ResponseWriter, r *http.Request) {
 		JSONError(w, "set_sudo_password", "Could not set sudo password", 500)
 		return
 	}
+
 	log.WithFields(log.Fields{
 		"event": "set_sudo_password",
 	}).Info("Successfully set sudo password.")
-	SimpleJSONResponse(w, "Successfully set '"+sudo_password+"' as sudo password. This password will not be exposed anywhere except inside the ./configs/config.json file. Make sure you don't commit this file to public repos :D", 200)
-}
 
-// @Summary      Get device control info
-// @Description  Provides the running containers, IOS devices info and apps available for installing
-// @Tags         devices
-// @Produce      json
-// @Success      200 {object} DeviceControlInfo
-// @Failure      500 {object} ErrorJSON
-// @Router       /devices/device-control [post]
-func GetDeviceControlInfo(w http.ResponseWriter, r *http.Request) {
-	var runningContainerNames = getRunningContainerNames()
-	var info = DeviceControlInfo{
-		RunningContainers: runningContainerNames,
-		IOSInfo:           getIOSDevicesInfo(runningContainerNames),
-		InstallableApps:   getInstallableApps(),
-		AndroidInfo:       getAndroidDevicesInfo(runningContainerNames),
-	}
-	fmt.Fprintf(w, PrettifyJSON(ConvertToJSONString(info)))
+	SimpleJSONResponse(w, "Successfully set '"+sudo_password+"' as sudo password. This password will not be exposed anywhere except inside the ./configs/config.json file. Make sure you don't commit this file to public repos :D", 200)
 }
 
 //=======================//
 //=====FUNCTIONS=====//
+
+// Completely setup udev and usbmuxd
+func SetupUdevListenerInternal() error {
+	DeleteTempUdevFiles()
+
+	err := CreateUdevRules()
+	if err != nil {
+		DeleteTempUdevFiles()
+		return err
+	}
+
+	err = SetUdevRules()
+	if err != nil {
+		DeleteTempUdevFiles()
+		return err
+	}
+
+	err = CopyFileShell("./configs/usbmuxd.service", "/lib/systemd/system/", sudo_password)
+	if err != nil {
+		DeleteTempUdevFiles()
+		return err
+	}
+
+	err = EnableUsbmuxdService()
+	if err != nil {
+		DeleteTempUdevFiles()
+		return err
+	}
+
+	DeleteTempUdevFiles()
+	return nil
+}
+
+// Remove udev rules for the devices
+func RemoveUdevListenerInternal() error {
+	// Delete the rules file
+	err := DeleteFileShell("/etc/udev/rules.d/90-device.rules", sudo_password)
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "remove_udev_listener",
+		}).Error("Could not delete udev rules file. Error: " + err.Error())
+		return err
+	}
+
+	// Reload udev after removing the rules
+	commandString := "echo '" + sudo_password + "' | sudo -S udevadm control --reload-rules"
+	cmd := exec.Command("bash", "-c", commandString)
+	err = cmd.Run()
+	if err != nil {
+		log.WithFields(log.Fields{
+			"event": "remove_udev_listener",
+		}).Error("Could not reload udev rules file. Error: " + err.Error())
+		return err
+	}
+
+	return nil
+}
 
 // Check if the sudo password in ./configs/config.json is different than "undefined" meaning something is set
 func CheckSudoPasswordSet() bool {
@@ -261,6 +365,8 @@ func CheckSudoPasswordSet() bool {
 		}).Error("Could not unmarshal ./configs/config.json file when checking sudo password")
 		return false
 	}
+
+	// Check the sudo password in the config
 	sudo_password := configData.EnvConfig.SudoPassword
 	if sudo_password == "undefined" || sudo_password == "" {
 		return false
@@ -294,7 +400,8 @@ func CreateUdevRules() error {
 	log.WithFields(log.Fields{
 		"event": "create_udev_rules",
 	}).Info("Creating udev rules")
-	// Create the rules file that will start usbmuxd on the first connected device
+
+	// Create the common devices udev rules file
 	create_container_rules, err := os.Create("./90-device.rules")
 	if err != nil {
 		return errors.New("Could not create 90-device.rules")
@@ -312,6 +419,7 @@ func CreateUdevRules() error {
 
 	devices_list := configData.DeviceConfig
 
+	// For each device generate the respective rule lines
 	for _, device := range devices_list {
 		rule_line1 := `SUBSYSTEM=="usb", ENV{ID_SERIAL_SHORT}=="` + device.DeviceUDID + `", MODE="0666", SYMLINK+="device_` + device.DeviceUDID + `"`
 		rule_line2 := `ACTION=="remove", ENV{ID_SERIAL_SHORT}=="` + device.DeviceUDID + `", RUN+="/usr/bin/curl -X POST -H \"Content-Type: application/json\" -d '{\"udid\":\"` + device.DeviceUDID + `\"}' http://localhost:10000/device-containers/remove"`
@@ -319,6 +427,7 @@ func CreateUdevRules() error {
 		//rule_line2 := `ACTION=="add", ENV{ID_SERIAL_SHORT}=="` + device.DeviceUdid + `", RUN+="/usr/local/bin/docker-cli start-device-container --device_type=` + device.OS + ` --udid=` + device.DeviceUdid + `"`
 		//rule_line3 := `ACTION=="remove", ENV{ID_SERIAL_SHORT}=="` + device.DeviceUdid + `", RUN+="/usr/local/bin/docker-cli remove-device-container --udid=` + device.DeviceUdid + `"`
 
+		// Write the new lines for each device in the udev rules file
 		if _, err := create_container_rules.WriteString(rule_line1 + "\n"); err != nil {
 			return errors.New("Could not write to 90-device.rules")
 		}
@@ -337,11 +446,13 @@ func CreateUdevRules() error {
 
 // Copy the iOS udev rules to /etc/udev/rules.d and reload udev
 func SetUdevRules() error {
-	//err := CopyFileShell("./39-usbmuxd.rules", "/etc/udev/rules.d/39-usbmuxd.rules", sudo_password)
+	// Copy the udev rules to /etc/udev/rules.d
 	err := CopyFileShell("./90-device.rules", "/etc/udev/rules.d/90-device.rules", sudo_password)
 	if err != nil {
 		return err
 	}
+
+	// Reload the udev rules after updating them
 	commandString := "echo '" + sudo_password + "' | sudo -S udevadm control --reload-rules"
 	cmd := exec.Command("bash", "-c", commandString)
 	err = cmd.Run()
@@ -363,7 +474,7 @@ func CheckWDAProvided() bool {
 	return true
 }
 
-// Get a value from ./configs/config.json
+// Get an env value from ./configs/config.json
 func GetEnvValue(key string) string {
 	configData, err := GetConfigJsonData()
 	if err != nil {
@@ -380,231 +491,4 @@ func GetEnvValue(key string) string {
 		return strconv.FormatBool(configData.EnvConfig.ConnectSeleniumGrid)
 	}
 	return ""
-}
-
-// Get the names of the currently running containers(that are for devices)
-func getRunningContainerNames() []string {
-	var containerNames []string
-
-	cli, err := client.NewClientWithOpts(client.FromEnv)
-	if err != nil {
-		return containerNames
-	}
-
-	// Get the current containers list
-	containers, err := cli.ContainerList(context.Background(), types.ContainerListOptions{All: true})
-	if err != nil {
-		return containerNames
-	}
-
-	// Loop through the containers list
-	for _, container := range containers {
-		// Parse plain container name
-		containerName := strings.Replace(container.Names[0], "/", "", -1)
-		if (strings.Contains(containerName, "iOSDevice") || strings.Contains(containerName, "androidDevice")) && strings.Contains(container.Status, "Up") {
-			containerNames = append(containerNames, containerName)
-		}
-	}
-	return containerNames
-}
-
-// For each running container extract the info for each respective device from ./configs/config.json to provide to the device-control info endpoint.
-// Provides installed apps, configuration info, wda urls
-func getIOSDevicesInfo(runningContainers []string) []IOSDeviceInfo {
-	var combinedInfo []IOSDeviceInfo
-	for _, containerName := range runningContainers {
-		if strings.Contains(containerName, "iOSDevice") {
-			// Extract the device UDID from the container name
-			re := regexp.MustCompile("[^_]*$")
-			device_udid := re.FindStringSubmatch(containerName)
-
-			var installed_apps []string
-			installed_apps, err := IOSDeviceApps(device_udid[0])
-			if err != nil {
-				installed_apps = append(installed_apps, "")
-			}
-
-			var device_config *IOSDevice
-			device_config, err = iOSDeviceConfig(device_udid[0])
-
-			var deviceInfo = IOSDeviceInfo{BundleIDs: installed_apps, DeviceConfig: device_config}
-			combinedInfo = append(combinedInfo, deviceInfo)
-		}
-	}
-	return combinedInfo
-}
-
-// For each running container extract the info for each respective device from ./configs/config.json to provide to the device-control info endpoint.
-// Provides installed apps, configuration info, wda urls
-func getAndroidDevicesInfo(runningContainers []string) []AndroidDeviceInfo {
-	var combinedInfo []AndroidDeviceInfo
-	for _, containerName := range runningContainers {
-		if strings.Contains(containerName, "androidDevice") {
-			// Extract the device UDID from the container name
-			re := regexp.MustCompile("[^_]*$")
-			device_udid := re.FindStringSubmatch(containerName)
-
-			var device_config *AndroidDevice
-			device_config, _ = androidDeviceConfig(device_udid[0])
-
-			var deviceInfo = AndroidDeviceInfo{DeviceConfig: device_config}
-			combinedInfo = append(combinedInfo, deviceInfo)
-		}
-	}
-	return combinedInfo
-}
-
-// Get the WDA and WDA stream urls from the container logs folder for a specific device
-func getIOSDeviceWdaURLs(device_udid string) (string, string, error) {
-	// Get the path of the WDA url file using regex
-	pattern := "./logs/*" + device_udid + "/ios-wda-url.json"
-	matches, err := filepath.Glob(pattern)
-
-	if err != nil {
-		return "", "", err
-	}
-
-	var wdaConfig WdaConfig
-	err = UnmarshalJSONFile(matches[0], &wdaConfig)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "get_wda_config",
-		}).Error("Could not unmarshal json at path: " + matches[0])
-		return "", "", err
-	}
-
-	return wdaConfig.WdaURL, wdaConfig.WdaStreamURL, nil
-}
-
-// Get the WDA and WDA stream urls from the container logs folder for a specific device
-func getAndroidDeviceMinicapStreamSize(device_udid string) (string, error) {
-	// Get the path of the WDA url file using regex
-	pattern := "./logs/*" + device_udid + "/minicap.log"
-	matches, err := filepath.Glob(pattern)
-
-	command_string := "cat " + matches[0] + " | grep \"+ args='-P\""
-	cmd := exec.Command("bash", "-c", command_string)
-	var out bytes.Buffer
-	cmd.Stdout = &out
-
-	err = cmd.Run()
-	if err != nil {
-		return "", err
-	}
-
-	if out.String() == "" {
-		log.WithFields(log.Fields{
-			"event": "get_minicap_stream_size",
-		}).Error("Error")
-		return "", nil
-	}
-
-	// Get the stream size that is between the -P and @ in the cmd out string
-	stream_size := GetStringInBetween(out.String(), "-P ", "@")
-	return stream_size, nil
-}
-
-// Get the configuration info for iOS device from ./configs/config.json
-func iOSDeviceConfig(device_udid string) (*IOSDevice, error) {
-	// Get the config data
-	configData, err := GetConfigJsonData()
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "ios_container_create",
-		}).Error("Could not unmarshal config.json file when trying to create a container for device with udid: " + device_udid)
-		return nil, err
-	}
-
-	// Check if device is registered in config data
-	var device_in_config bool
-	var deviceConfig DeviceConfig
-	for _, v := range configData.DeviceConfig {
-		if v.DeviceUDID == device_udid {
-			device_in_config = true
-			deviceConfig = v
-		}
-	}
-
-	// Stop execution if device not in config data
-	if !device_in_config {
-		log.WithFields(log.Fields{
-			"event": "android_container_create",
-		}).Error("Device with UDID:" + device_udid + " is not registered in the './configs/config.json' file. No container will be created.")
-		return nil, err
-	}
-
-	wda_url, wda_stream_url, err := getIOSDeviceWdaURLs(device_udid)
-	if err != nil {
-		return nil, err
-	}
-
-	return &IOSDevice{
-			AppiumPort:         deviceConfig.AppiumPort,
-			DeviceName:         deviceConfig.DeviceName,
-			DeviceOSVersion:    deviceConfig.DeviceOSVersion,
-			WdaMjpegPort:       deviceConfig.WDAMjpegPort,
-			WdaPort:            deviceConfig.WDAPort,
-			DeviceUDID:         device_udid,
-			WdaMjpegURL:        wda_stream_url,
-			WdaURL:             wda_url,
-			DeviceModel:        "Remove please",
-			DeviceViewportSize: deviceConfig.ViewportSize},
-		nil
-}
-
-// Get the configuration info for iOS device from ./configs/config.json
-func androidDeviceConfig(device_udid string) (*AndroidDevice, error) {
-	// Get the config data
-	configData, err := GetConfigJsonData()
-	if err != nil {
-		log.WithFields(log.Fields{
-			"event": "android_container_create",
-		}).Error("Could not unmarshal config.json file when trying to create a container for device with udid: " + device_udid)
-		return nil, err
-	}
-
-	// Check if device is registered in config data
-	var device_in_config bool
-	var deviceConfig DeviceConfig
-	for _, v := range configData.DeviceConfig {
-		if v.DeviceUDID == device_udid {
-			device_in_config = true
-			deviceConfig = v
-		}
-	}
-
-	// Stop execution if device not in config data
-	if !device_in_config {
-		log.WithFields(log.Fields{
-			"event": "android_container_create",
-		}).Error("Device with UDID:" + device_udid + " is not registered in the './configs/config.json' file. No container will be created.")
-		return nil, err
-	}
-
-	stream_size, err := getAndroidDeviceMinicapStreamSize(device_udid)
-	return &AndroidDevice{
-			AppiumPort:      deviceConfig.AppiumPort,
-			DeviceName:      deviceConfig.DeviceName,
-			DeviceOSVersion: deviceConfig.DeviceOSVersion,
-			DeviceUDID:      device_udid,
-			StreamSize:      stream_size,
-			StreamPort:      strconv.Itoa(deviceConfig.StreamPort)},
-		nil
-}
-
-// Get the installable apps list from the ./apps folder
-func getInstallableApps() []string {
-	var appNames []string
-
-	files, err := ioutil.ReadDir("./apps/")
-	if err != nil {
-		return appNames
-	}
-
-	for _, file := range files {
-		if strings.Contains(file.Name(), ".ipa") || strings.Contains(file.Name(), ".app") || strings.Contains(file.Name(), ".apk") {
-			appNames = append(appNames, file.Name())
-		}
-	}
-	return appNames
 }
