+++ static/device_containers.html
@@ -3,7 +3,7 @@
 
 <head>
 	<meta charset="UTF-8">
-	<title>iOS Containers Info</title>
+	<title>Device Containers Info</title>
 	<link href="/static/css/device_containers_styles.css" rel="stylesheet">
 	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 	<script src="https://unpkg.com/sweetalert/dist/sweetalert.min.js"></script>
