+++ static/device_control.html
@@ -285,7 +285,7 @@
         let selected_device_text = $("#device-dropdown option:selected").text()
         let table_element = $("#element-info-table")
 
-        if (selected_device_text.includes("iOS")) {
+        if (selected_device_text.includes("ios")) {
             table_element.empty()
             table_element.append('<tr><th>type</th><td id="element-type"></td></tr>')
             table_element.append('<tr><th>name</th><td id="element-name"></td></tr>')
@@ -329,7 +329,7 @@
         if (selected_device_text.includes("android")) {
             url = "http://127.0.0.1:" + $('#appium-port').text() + "/wd/hub/session/" + $("#appium-session-id").text() + "/appium/device/press_keycode"
             json = JSON.stringify({ "keycode": 3 })
-        } else if (selected_device_text.includes("iOS")) {
+        } else if (selected_device_text.includes("ios")) {
             let wda_url = $("#wda-url").text()
             url = wda_url + "/wda/homescreen"
         } else {
@@ -356,7 +356,7 @@
 
         if (selected_device_text.includes("android")) {
             url = "http://127.0.0.1:" + $('#appium-port').text() + "/wd/hub/session/" + $("#appium-session-id").text() + "/appium/device/unlock"
-        } else if (selected_device_text.includes("iOS")) {
+        } else if (selected_device_text.includes("ios")) {
             let wda_url = $("#wda-url").text()
             let wda_session_id = $("#wda-session-id").text()
             url = wda_url + "/session/" + wda_session_id + "/wda/unlock"
@@ -431,7 +431,7 @@
 
         if (selected_device_text.includes("android")) {
             active_element_url = "http://127.0.0.1:" + $('#appium-port').text() + "/wd/hub/session/" + $("#appium-session-id").text() + "/element/active"
-        } else if (selected_device_text.includes("iOS")) {
+        } else if (selected_device_text.includes("ios")) {
             wda_url = $("#wda-url").text()
             wda_session_id = $("#wda-session-id").text()
             active_element_url = wda_url + "/session/" + wda_session_id + "/element/active"
@@ -477,7 +477,7 @@
 
         if (selected_device_text.includes("android")) {
             active_element_url = "http://127.0.0.1:" + $('#appium-port').text() + "/wd/hub/session/" + $("#appium-session-id").text() + "/element/active"
-        } else if (selected_device_text.includes("iOS")) {
+        } else if (selected_device_text.includes("ios")) {
             wda_url = $("#wda-url").text()
             wda_session_id = $("#wda-session-id").text()
             active_element_url = wda_url + "/session/" + wda_session_id + "/element/active"
@@ -527,7 +527,7 @@
             let device_udid = $("#device-dropdown").val();
             let selected_device_text = $("#device-dropdown option:selected").text()
 
-            if (selected_device_text.includes("iOS")) {
+            if (selected_device_text.includes("ios")) {
                 let wda_url;
                 for (let i = 0; i < control_response["ios-devices-info"].length; i++) {
                     if (control_response["ios-devices-info"][i].deviceConfig.device_udid === device_udid) {
@@ -672,7 +672,7 @@
                 table_element.append('<tr><th>WDA Mjpeg Port:</th><td>' + response["ios-devices-info"][i].deviceConfig.wda_mjpeg_port + '</td></tr>')
                 table_element.append('<tr><th>WDA URL:</th><td id="wda-url">' + response["ios-devices-info"][i].deviceConfig.wda_url + '</td></tr>')
                 table_element.append('<tr><th>WDA Stream URL:</th><td>' + response["ios-devices-info"][i].deviceConfig.wda_stream_url + '</td></tr>')
-                table_element.append('<tr><th>Device screen size:</th><td id="device-size">' + response["ios-devices-info"][i].deviceConfig.viewport_size + '</td></tr>')
+                table_element.append('<tr><th>Device screen size:</th><td id="device-size">' + response["ios-devices-info"][i].deviceConfig.screen_size + '</td></tr>')
                 table_element.append('<tr><th>WDA session ID:</th><td id="wda-session-id">No session</td></tr>')
             }
         }
@@ -780,7 +780,7 @@
 
         //var response = JSON.parse(data)
         for (let i = 0; i < response["installable-apps"].length; i++) {
-            if (container_name.includes("iOSDevice") && response["installable-apps"][i].includes(".ipa") || response["installable-apps"][i].includes(".app")) {
+            if (container_name.includes("iosDevice") && response["installable-apps"][i].includes(".ipa") || response["installable-apps"][i].includes(".app")) {
                 apps_dropdown.append($('<option></option>').attr('value', response["installable-apps"][i]).text(response["installable-apps"][i]));
             } else if (container_name.includes("androidDevice") && response["installable-apps"][i].includes(".apk")) {
                 apps_dropdown.append($('<option></option>').attr('value', response["installable-apps"][i]).text(response["installable-apps"][i]));
@@ -794,7 +794,7 @@
         let image_stream = $('#image-stream')
         let selected_device_text = $("#device-dropdown option:selected").text()
 
-        if (selected_device_text.includes("iOS")) {
+        if (selected_device_text.includes("ios")) {
             for (let i = 0; i < response["ios-devices-info"].length; i++) {
                 if (response["ios-devices-info"][i].deviceConfig.device_udid === device_udid) {
                     image_stream.attr('src', response["ios-devices-info"][i].deviceConfig.wda_stream_url)
@@ -933,7 +933,7 @@
 
         if (selected_device_text.includes("android")) {
             url = "http://127.0.0.1:" + $('#appium-port').text() + "/wd/hub/session/" + $("#appium-session-id").text() + "/source"
-        } else if (selected_device_text.includes("iOS")) {
+        } else if (selected_device_text.includes("ios")) {
             wda_url = $("#wda-url").text()
             wda_session_id = $("#wda-session-id").text()
             url = wda_url + "/session/" + wda_session_id + "/source"
@@ -1125,7 +1125,7 @@
     function showElementInfo(json) {
         let selected_device_text = $("#device-dropdown option:selected").text()
 
-        if (selected_device_text.includes("iOS")) {
+        if (selected_device_text.includes("ios")) {
             $("#element-info-table").css("visibility", "visible")
             $("#element-type").text(json.type)
             $("#element-name").text(json.name)
@@ -1405,7 +1405,7 @@
         let appium_height;
 
         let selected_device_text = $("#device-dropdown option:selected").text()
-        if (selected_device_text.includes("iOS")) {
+        if (selected_device_text.includes("ios")) {
             // get the selected element coordinates and width/height from the json
             elementX = json.x;
             elementY = json.y;
