+++ static/project_config.html
@@ -111,14 +111,14 @@
                 </th>
             </tr>
             <tr>
-                <td>{{.ProjectConfigValues.DevicesHost}}</td>
-                <td>{{.ProjectConfigValues.SeleniumHubHost}}</td>
-                <td>{{.ProjectConfigValues.SeleniumHubPort}}</td>
-                <td>{{.ProjectConfigValues.SeleniumHubProtocolType}}</td>
-                <td>{{.ProjectConfigValues.WDABundleID}}</td>
+                <td>{{.AppiumConfigValues.DevicesHost}}</td>
+                <td>{{.AppiumConfigValues.SeleniumHubHost}}</td>
+                <td>{{.AppiumConfigValues.SeleniumHubPort}}</td>
+                <td>{{.AppiumConfigValues.SeleniumHubProtocolType}}</td>
+                <td>{{.AppiumConfigValues.WDABundleID}}</td>
                 <td id="wda-status-cell">{{.WebDriverAgentProvided}}</td>
                 <td id="image-status-cell">{{.ImageStatus}}</td>
-                <td id="udev-ios-listener-status-cell">{{.UdevIOSListenerStatus}}</td>
+                <td id="udev-udev-listener-status-cell">{{.UdevIOSListenerStatus}}</td>
                 <td id="sudo-password-cell">{{.SudoPasswordSet}}</td>
             </tr>
         </table>
